+++ src/main/java/walkingkooka/spreadsheet/server/SpreadsheetHttpServerApiSpreadsheetEngineHttpHandlerRequest.java
@@ -23,7 +23,7 @@ import walkingkooka.net.http.HttpStatusCode;
 import walkingkooka.net.http.server.HttpRequest;
 import walkingkooka.net.http.server.HttpRequestAttributes;
 import walkingkooka.net.http.server.HttpResponse;
-import walkingkooka.spreadsheet.SpreadsheetId;
+import walkingkooka.spreadsheet.meta.SpreadsheetId;
 
 import java.util.Objects;
 import java.util.Optional;
