+++ src/main/java/walkingkooka/spreadsheet/server/meta/SpreadsheetMetadataHateosResourceHandler.java
@@ -18,7 +18,7 @@
 package walkingkooka.spreadsheet.server.meta;
 
 import walkingkooka.net.http.server.hateos.HateosResourceHandler;
-import walkingkooka.spreadsheet.SpreadsheetId;
+import walkingkooka.spreadsheet.meta.SpreadsheetId;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadata;
 import walkingkooka.store.Store;
