+++ src/main/java/walkingkooka/spreadsheet/server/meta/SpreadsheetMetadataHttpHandler.java
@@ -31,7 +31,7 @@ import walkingkooka.net.http.server.HttpResponse;
 import walkingkooka.net.http.server.hateos.HateosResourceMappings;
 import walkingkooka.route.RouteMappings;
 import walkingkooka.route.Router;
-import walkingkooka.spreadsheet.SpreadsheetId;
+import walkingkooka.spreadsheet.meta.SpreadsheetId;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadata;
 import walkingkooka.spreadsheet.server.SpreadsheetHttpServer;
 import walkingkooka.spreadsheet.server.SpreadsheetServerContext;
