+++ src/main/java/walkingkooka/spreadsheet/server/net/SpreadsheetUrlPathTemplate.java
@@ -19,9 +19,9 @@ package walkingkooka.spreadsheet.server.net;
 
 import walkingkooka.collect.map.Maps;
 import walkingkooka.net.UrlPath;
-import walkingkooka.spreadsheet.SpreadsheetId;
 import walkingkooka.spreadsheet.engine.SpreadsheetEngineEvaluation;
 import walkingkooka.spreadsheet.format.provider.SpreadsheetFormatterSelector;
+import walkingkooka.spreadsheet.meta.SpreadsheetId;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadataPropertyName;
 import walkingkooka.spreadsheet.meta.SpreadsheetName;
 import walkingkooka.spreadsheet.reference.SpreadsheetColumnReferenceOrRange;
