+++ src/test/java/walkingkooka/spreadsheet/server/meta/SpreadsheetMetadataHateosResourceHandlerLoadTest.java
@@ -25,7 +25,7 @@ import walkingkooka.collect.set.SortedSets;
 import walkingkooka.environment.AuditInfo;
 import walkingkooka.net.UrlPath;
 import walkingkooka.net.http.server.hateos.HateosResourceHandler;
-import walkingkooka.spreadsheet.SpreadsheetId;
+import walkingkooka.spreadsheet.meta.SpreadsheetId;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadata;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadataPropertyName;
 import walkingkooka.spreadsheet.meta.SpreadsheetName;
