+++ src/test/java/walkingkooka/spreadsheet/server/meta/SpreadsheetMetadataSetTest.java
@@ -22,7 +22,7 @@ import walkingkooka.collect.set.Sets;
 import walkingkooka.net.http.server.hateos.HateosResourceSetTesting;
 import walkingkooka.reflect.ClassTesting;
 import walkingkooka.reflect.JavaVisibility;
-import walkingkooka.spreadsheet.SpreadsheetId;
+import walkingkooka.spreadsheet.meta.SpreadsheetId;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadata;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadataPropertyName;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadataTesting;
